refactor: jobWatcher.run
We shouldn't need to error out if we can't get the allocation that we
just found - if we can't, it likely died between those times and we can
just wait 5 seconds again.

Thus, we can range over time.Ticker and continue any time we hit an
"error" in the loop.

Lastly, removes the stray "\n"s from the end of the log statements; log
is implicitly line-oriented so they aren't needed.